// web/src/graphql-types.d.ts
export interface PageInfo {
  hasNextPage: boolean;
}
// ...
export interface Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
}
// ...
export type Date = any;
// ...
export interface File extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  childAuthorJson: AuthorJson | null;
  childImageSharp: ImageSharp | null;
  childMarkdownRemark: MarkdownRemark | null;
  internal: internal_12 | null;
  sourceInstanceName: string | null;
  absolutePath: string | null;
  relativePath: string | null;
  extension: string | null;
  size: number | null;
  prettySize: string | null;
  modifiedTime: Date | null;
  accessTime: Date | null;
  changeTime: Date | null;
  birthTime: Date | null;
  root: string | null;
  dir: string | null;
  base: string | null;
  ext: string | null;
  name: string | null;
  relativeDirectory: string | null;
  dev: number | null;
  mode: number | null;
  nlink: number | null;
  uid: number | null;
  gid: number | null;
  rdev: number | null;
  blksize: number | null;
  ino: number | null;
  blocks: number | null;
  atimeMs: number | null;
  mtimeMs: number | null;
  ctimeMs: number | null;
  birthtimeMs: number | null;
  atime: Date | null;
  mtime: Date | null;
  ctime: Date | null;
  birthtime: Date | null;
  publicURL: string | null;
}
// ...
export interface AuthorJson extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  bio: string | null;
  avatar: File | null;
  twitter: string | null;
  github: string | null;
  internal: internal_13 | null;
}

export interface internal_13 {
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
}

export interface ImageSharp extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  internal: internal_14 | null;
  fixed: ImageSharpFixed | null;
  resolutions: ImageSharpResolutions | null;
  fluid: ImageSharpFluid | null;
  sizes: ImageSharpSizes | null;
  original: ImageSharpOriginal | null;
  resize: ImageSharpResize | null;
}
// ...
export interface internal_14 {
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
}
// ...
export interface ImageSharpFixed {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  width: number | null;
  height: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  originalName: string | null;
}

export interface ImageSharpResolutions {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  width: number | null;
  height: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  originalName: string | null;
}

export interface ImageSharpFluid {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  sizes: string | null;
  originalImg: string | null;
  originalName: string | null;
}

export interface ImageSharpSizes {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  sizes: string | null;
  originalImg: string | null;
  originalName: string | null;
}

export interface ImageSharpOriginal {
  width: number | null;
  height: number | null;
  src: string | null;
}

export interface ImageSharpResize {
  src: string | null;
  tracedSVG: string | null;
  width: number | null;
  height: number | null;
  aspectRatio: number | null;
  originalName: string | null;
}

export interface MarkdownRemark extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  internal: internal_15 | null;
  frontmatter: frontmatter_2 | null;
  rawMarkdownBody: string | null;
  fileAbsolutePath: string | null;
  fields: fields_2 | null;
  html: string | null;
  htmlAst: JSON | null;
  excerpt: string | null;
  headings: Array<MarkdownHeading> | null;
  timeToRead: number | null;
  tableOfContents: string | null;
  wordCount: wordCount | null;
}

export interface ExcerptMarkdownRemarkArgs {
  pruneLength: number | null;
  truncate: boolean | null;
}

export interface HeadingsMarkdownRemarkArgs {
  depth: HeadingLevels | null;
}

export interface internal_15 {
  content: string | null;
  type: string | null;
  contentDigest: string | null;
  owner: string | null;
  fieldOwners: fieldOwners_2 | null;
}

export interface fieldOwners_2 {
  slug: string | null;
}

export interface frontmatter_2 {
  title: string | null;
  createdDate: Date | null;
  updatedDate: Date | null;
  author: AuthorJson | null;
  tags: Array<string> | null;
  image: File | null;
  draft: boolean | null;
  _PARENT: string | null;
}

export interface CreatedDatefrontmatter_2Args {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface UpdatedDatefrontmatter_2Args {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface fields_2 {
  slug: string | null;
}

export type JSON = any;

export type HeadingLevels = "h1" | "h2" | "h3" | "h4" | "h5" | "h6";

export interface MarkdownHeading {
  value: string | null;
  depth: number | null;
}

export interface wordCount {
  paragraphs: number | null;
  sentences: number | null;
  words: number | null;
}
// ...
export interface MarkdownRemarkConnection {
  pageInfo: PageInfo;
  edges: Array<MarkdownRemarkEdge> | null;
  totalCount: number | null;
  distinct: Array<string> | null;
  group: Array<markdownRemarkGroupConnectionConnection> | null;
}

export interface DistinctMarkdownRemarkConnectionArgs {
  field: markdownRemarkDistinctEnum | null;
}

export interface GroupMarkdownRemarkConnectionArgs {
  skip: number | null;
  limit: number | null;
  field: markdownRemarkGroupEnum | null;
}

export interface MarkdownRemarkEdge {
  node: MarkdownRemark | null;
  next: MarkdownRemark | null;
  previous: MarkdownRemark | null;
}
// ...
export interface markdownRemarkGroupConnectionConnection {
  pageInfo: PageInfo;
  edges: Array<markdownRemarkGroupConnectionEdge> | null;
  field: string | null;
  fieldValue: string | null;
  totalCount: number | null;
}

export interface markdownRemarkGroupConnectionEdge {
  node: MarkdownRemark | null;
  next: MarkdownRemark | null;
  previous: MarkdownRemark | null;
}
// ...
export interface Site extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  siteMetadata: siteMetadata_2 | null;
  port: Date | null;
  host: string | null;
  mapping: mapping_2 | null;
  pathPrefix: string | null;
  polyfill: boolean | null;
  buildTime: Date | null;
  internal: internal_16 | null;
}

export interface PortSiteArgs {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface BuildTimeSiteArgs {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface siteMetadata_2 {
  title: string | null;
  googleVerification: string | null;
  disqus: string | null;
}

export interface mapping_2 {
  MarkdownRemark_frontmatter_author: string | null;
}

export interface internal_16 {
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
}